% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.order.R
\name{col.order}
\alias{col.order}
\title{Sorted by descending mean for each column}
\usage{
col.order(data, group = 1)
}
\arguments{
\item{data}{a data frame containing the grouping information in the first column}

\item{group}{NA or a numeric value indicate which variable containing the grouping inf.}
}
\value{
a ordered data frame
}
\description{
Sorted by descending mean for each column
}
\examples{
col.order(data = iris, group = 5)
col.order(data = iris[1:4], group = NA)
}
\author{
Zhonghui Gai
}
