% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{createTable}
\alias{createTable}
\title{create Groups}
\usage{
createTable(
  x,
  hide = NA,
  digits = NA,
  type = NA,
  show.p.overall = TRUE,
  show.all,
  show.p.trend,
  show.p.mul = FALSE,
  show.n,
  show.ratio = FALSE,
  show.descr = TRUE,
  show.ci = FALSE,
  hide.no = NA,
  digits.ratio = NA,
  show.p.ratio = show.ratio,
  digits.p = 3,
  sd.type = 1,
  q.type = c(1, 1),
  extra.labels = NA,
  all.last = FALSE
)
}
\description{
create Groups
}
