% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groutable.R
\name{groutable}
\alias{groutable}
\title{This is a wrapper function of compareGroups package}
\usage{
groutable(data, variable = 1, method = 1, writeTab = FALSE, digits = 3)
}
\arguments{
\item{data}{a data frame}

\item{variable}{A numeric to indicate which variable to be used for grouping}

\item{method}{1 - forces analysis as "normal-distributed";2 - forces analysis as "continuous non-normal";
3 - forces analysis as "categorical";4 - NA, which performs a Shapiro-Wilks test to decide between normal or non-normal.}

\item{writeTab}{A boolean value to indicate whether to save the result or not}

\item{digits}{show the number of digits}
}
\value{
a statistical table
}
\description{
This is a wrapper function of compareGroups package
}
\examples{
groutable(data = iris, variable = 5)
groutable(data = iris[1:100,], variable = 5)
groutable(data = droplevels(iris[1:100,]), variable = 5)
groutable(data = droplevels(iris[1:100,]), variable = 5, digits = 2)
groutable(data = droplevels(iris[1:100,]), variable = 5, digits = 2, method = 4)
data <- iris
data$Species <- as.character(data$Species)
groutable(data = droplevels(data[1:100,]), variable = 5, digits = 2, method = 4)
}
\author{
Zhonghui Gai
}
